requirements.txt
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==4.2.9
django-cors-headers==4.3.1
django-filter==24.1
django-rest-framework-docs==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
inflection==0.5.1
jsonpickle==3.0.2
numpy==1.24.4
openai==1.7.2
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
PyKakao==0.0.7
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
wfastcgi==3.0.0
xmltodict==0.13.0
